requirements.txt
Django==1.10.5
Pillow
psycopg2==2.6.2
django-crispy-forms==1.6.1
python-dateutil
django-registration-redux==1.4
social-auth-app-django==1.1.0
django-braces==1.11.0
djangorestframework==3.6.2